backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-google-genai==0.0.6
python-dotenv==1.0.0
